♻️refactor: GamePanel 분리
- 추후에 새로운 게임 추가되면 그에 맞춰 변경 될 예정
- UserList section 디자인 이관
