+++ src/components/BalanceGameProgress/BalanceGameProgress.tsx
@@ -32,7 +32,10 @@ const BalanceGameProgress = ({
 
   useEffect(() => {
     if (isTimeout || isAllSelected) {
+      console.log('시간초과?', isTimeout);
+      console.log('전부선택?', isAllSelected);
       setRoomStatus('result');
+      resetSelectedPlayers();
 
       getBalanceGameResults({
         roomId: roomId,
@@ -46,8 +49,6 @@ const BalanceGameProgress = ({
         .catch((error) => {
           console.log(error);
         });
-
-      resetSelectedPlayers();
     }
   }, [isTimeout, isAllSelected]);
